Change TournamentOptions to be made by builders
Started with DoubleEliminationsOptionsBuilder.  Since all of the tournament options have default values, it makes more sense to have a builder for each type of tournament options that generates a single TournamentOptions object.

+++ challonge-api-v2/src/main/java/ChallongeClient.java
@@ -45,11 +45,11 @@ public class ChallongeClient {
         JSONObject response = this.api.apiGet(ChallongeApi.toURI("tournaments"));
         JSONObject meta =
         TypeUtils.requireType(response, "meta", JSONObject.class);
-        long count =
-        TypeUtils.requireType(meta, "count", Long.class);
+        int count =
+        (int)(long)TypeUtils.requireType(meta, "count", Long.class);
 
         ChallongeTournament[] tournaments =
-        new ChallongeTournament[(int)count];
+        new ChallongeTournament[count];
         int i = 0;
         while (true) {
             JSONArray data =
